[BTAT-2248] Refactor models.CustomReads to be part of the models.core package
